# avinya-hakathon RAG(2)/medical_knowledge/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Readout Medical Intelligence & Clinical Decision-Support API",
    "description": "Comprehensive full-stack API for laboratory report OCR extraction, unit normalization, reference range evaluation, clinical pattern detection, AI decision support, and longitudinal trend analysis.",
    "version": "2.0.0",
    "contact": {
      "name": "Readout Clinical Support",
      "email": "clinical-support@readout.health"
    }
  },
  "servers": [
    {
      "url": "/api",
      "description": "Readout Active Server"
    }
  ],
...
  "paths": {
...
    "/auth/register": {
      "post": {
        "tags": ["Auth"],
        "summary": "Register Patient or Clinician",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": { "type": "string", "example": "patient@example.com" },
                  "password": { "type": "string", "example": "SecurePass123!" },
                  "fullName": { "type": "string", "example": "Ananya Sharma" }
                },
                "required": ["email", "password"]
              }
            }
          }
        },
        "responses": { "200": { "description": "Registration successful." } }
      }
    },
    "/auth/login": {
      "post": {
        "tags": ["Auth"],
        "summary": "User Login",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": { "type": "string", "example": "demo@readout.health" },
                  "password": { "type": "string", "example": "demopassword123" }
                },
                "required": ["email", "password"]
              }
            }
          }
        },
        "responses": { "200": { "description": "Login successful, returns JWT token." } }
      }
    },
...
    "/reports/upload": {
      "post": {
        "tags": ["Reports"],
        "summary": "Upload Laboratory Report",
        "description": "Accepts PDF or image of medical report, runs OCR layout analysis, and extracts parameters.",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "file": { "type": "string", "format": "binary" },
                  "fasting": { "type": "boolean", "default": true },
                  "title": { "type": "string", "example": "Annual Comprehensive Health Panel" }
                }
              }
            }
          }
        },
        "responses": { "200": { "description": "Report accepted and queued for extraction." } }
      }
    },
...
    "/reports/{id}/status": {
      "get": {
        "tags": ["Reports"],
        "summary": "Get Processing Status",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": { "200": { "description": "Processing status (UPLOADED, OCR, EXTRACTING, COMPLETED)." } }
      }
    },
...
    "/medications/{name}": {
      "get": {
        "tags": ["Medications"],
        "summary": "Educational Medication Reference",
        "parameters": [{ "name": "name", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": {
          "200": {
            "description": "Medication educational data with label: 'MEDICATION INFORMATION — FOR PROFESSIONAL REVIEW'."
          }
        }
      }
    },
...
    "/chat": {
      "post": {
        "tags": ["Chat"],
        "summary": "Ask My Reports (Grounded Lab Chat)",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "message": { "type": "string", "example": "What was my fasting glucose reading?" }
                },
                "required": ["message"]
              }
            }
          }
        },
        "responses": { "200": { "description": "Evidence-grounded response with source citations." } }
      }
    },
...
  }
}